requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
pytest==8.3.5
SQLAlchemy==2.0.40
uvicorn==0.34.2